apps/documents/requirements.txt
pymongo
kafka-python
pillow
pdf2image
pytesseract
pypdf2
opencv-python
grpcio
grpcio-tools
protobuf
pytest
insightface
onnxruntime